Move admin auth middleware to shared middleware package
